+++ Sources/SwiftUIKit/Views/CircularProgressView.swift
@@ -13,20 +13,18 @@ import SwiftUI
  This view is just a convenience for a `ProgressView` with a
  `CircularProgressViewStyle` applied.
  */
+@available(iOS 14.0, tvOS 14.0, watchOS 7.0, *)
 public struct CircularProgressView: View {
     
     public init() {}
     
     public var body: some View {
-        if #available(iOS 14.0, tvOS 14.0, watchOS 7.0, *) {
-            ProgressView()
-                .progressViewStyle(CircularProgressViewStyle())
-        } else {
-            // Fallback on earlier versions
-        }
+        ProgressView()
+            .progressViewStyle(.circular)
     }
 }
 
+@available(iOS 14.0, tvOS 14.0, watchOS 7.0, *)
 struct CircularProgressView_Previews: PreviewProvider {
     
     static var previews: some View {