+++ Sources/SwiftUIKit/Views/List/ListTextContent.swift
@@ -1,9 +1,9 @@
 //
-//  ListText.swift
+//  ListTextContent.swift
 //  SwiftUIKit
 //
-//  Created by Daniel Saidi on 2021-10-28.
-//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//  Created by Daniel Saidi on 2022-02-04.
+//  Copyright © 2022 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -12,12 +12,12 @@ import SwiftUI
  This view embeds the provided texts and icon in an `HStack`
  and wraps the stack in a ``ListItem``.
  
- Use this view when you want to render the content as a full
- list item, with all the padding and other tweaks that comes
- with it. Use ``ListTextContent`` to just get plain content.
+ Use this view when you just want the plain content, without
+ any list item modifications. Use ``ListText`` to get a full
+ list item rendering of the same content.
  */
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public struct ListText<Icon: View>: View {
+public struct ListTextContent<Icon: View>: View {
     
     public init(
         _ text: String,
@@ -33,14 +33,23 @@ public struct ListText<Icon: View>: View {
     private let subtitle: String?
     
     public var body: some View {
-        ListItem {
-            ListTextContent(text, icon: icon, subtitle: subtitle)
+        HStack {
+            if let icon = icon {
+                Label { Text(text) } icon: { icon }
+            } else {
+                Text(text)
+            }
+            
+            if let subtitle = subtitle {
+                Spacer()
+                ListSubtitle(subtitle)
+            }
         }
     }
 }
 
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ListText where Icon == EmptyView {
+public extension ListTextContent where Icon == EmptyView {
     
     init(
         _ text: String,
@@ -52,7 +61,7 @@ public extension ListText where Icon == EmptyView {
 }
 
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ListText where Icon == Image {
+public extension ListTextContent where Icon == Image {
     
     init(
         _ text: String,
@@ -65,19 +74,19 @@ public extension ListText where Icon == Image {
 }
 
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-struct ListText_Previews: PreviewProvider {
+struct ListTextContent_Previews: PreviewProvider {
     
     static var previews: some View {
         List {
-            ListText("Text 1")
-            ListText("Text 2")
-            ListText(
+            ListTextContent("Text 1")
+            ListTextContent("Text 2")
+            ListTextContent(
                 "Text with icon",
                 icon: Image(systemName: "lightbulb"))
-            ListText(
+            ListTextContent(
                 "Text with color badge",
                 icon: Color.blue.clipShape(Circle()).padding(1))
-            ListText(
+            ListTextContent(
                 "Text with icon and subtitle",
                 icon: Image(systemName: "lightbulb"),
                 subtitle: "Subtitle")