+++ Sources/SwiftUIKit/_Deprecated/ListItem+Deprecated.swift
@@ -14,6 +14,7 @@ public typealias ListButtonLinkItem = ListButtonLink
 @available(*, deprecated, renamed: "ListNavigationLink")
 public typealias ListNavigationLinkItem = ListNavigationLink
 
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
 @available(*, deprecated, renamed: "ListText")
 public typealias ListTextItem = ListText